removed files, cleaned up directories.
