--- lambda-functions/RDS/README.md
@@ -1,6 +1,0 @@
-# RDS
-
-In this repo all of the RDS related Lambda functions will reside here
-
-**DISCLAIMER**
-- Files in this repository are not intended to be tested locally or on any test environment, just with AWS tools and services.
