+++ lambda-functions/src/RDS/db_utils.py
@@ -1,30 +1,30 @@
 import psycopg2
+import logging
+import sys
 from subprocess import call
-
-# as the name implies, access the RDS database with pyscopg2 without the need of an ORM like SQLAlchemy
-# DBInstanceIdentifier can be found by running the command after the ` in your shell
-# this is a dependency file of our lambda function postgres_makeconn.py
 
 db_host = call("aws rds describe-db-instances | jq -r '.DBInstances[]|select(.DBInstanceIdentifier=='rds-reddit-data').Endpoint|.Address'", shell=True)
 db_port = 5432
 db_name = "hiddenalphabet_reddit_data"
 db_user = "hiddenalphabet"
 db_pass = "hiddenalphabet"
-db_table = "reddit-comments"
+db_table = "?"
 
 def make_conn():
+	
     conn = None
     try:
         conn = psycopg2.connect("dbname='%s' user='%s' host='%s' password='%s'" % (db_name,db_user,db_host,db_pass))
     except:
-        print('Unable to connect to Database.')
+        logger.error("Unexpected error: Could not connect to RDS-PostgreSQL instance")
+        sys.exit()
 
     return conn
 
 def fetch(conn,query):
-    result = []
-    print('Executing Query : %s' % (query))
 
+    print('Executing Query : %s' % (query))
+    result = []
     cursor = conn.cursor()
     cursor.execute(query)
     raw = cursor.fetchall()
@@ -33,5 +33,3 @@ def fetch(conn,query):
         result.append(line)
 
     return result
-
-# Lambda function logic for accessing the PostgreSQL DB in RDS
