+++ lambda-functions/src/SNS/print_SNS_message.py
